python/requirement.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn==0.27.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email validation
email-validator==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1